+ new files

+++ src/main/java/by/mordas/project/dao/AbstractDAO.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface AbstractDAO<T extends Entity>{
     public abstract List<T> findAllEntity() throws DAOException;
-    public abstract T findEntityById(int id) throws DAOException;
-    public abstract boolean delete(int id) throws DAOException;
+    public abstract T findEntityById(long id) throws DAOException;
+    public abstract boolean delete(long id) throws DAOException;
     public abstract void create(T entity) throws DAOException;
     public abstract T update(T entity) throws DAOException;
 }
