+++ src/main/java/by/mordas/project/dao/FacultyDAO.java
@@ -16,9 +16,9 @@ public interface FacultyDAO extends AbstractDAO<Faculty> {
 
     List<Faculty> findAllEntity() throws DAOException;
 
-    Faculty findEntityById(int id) throws DAOException;
+    Faculty findEntityById(long id) throws DAOException;
 
-    boolean delete(int id) throws DAOException;
+    boolean delete(long id) throws DAOException;
 
     void create(Faculty faculty) throws DAOException;
